+++ NEWS.md
@@ -5,6 +5,8 @@ User-visible changes worth mentioning.
 - [#823] Make configuration and specs ORM independent
 - [#777] Add support for public client in password grant flow
 - [#745] Add created_at timestamp to token generation options
+- [#838] Drop `Application#scopes` generator and warning, introduced for
+  upgrading doorkeeper from v2 to v3.
 
 ---